pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
[tool.poetry.dependencies]
aiohttp = "*"
python = "^3.7"
python-socketio = "*"
flore1 = "^1.5"
